Adjusted CMake paths to use MICROS_ROS_FIRMWARE_DIR
This should not change any functionality for the apps, but allow them to
be used as a templates when starting with external custom app folders
(introduced in micro-ROS/micro_ros_setup#183). The currently used
CMAKE_CURRENT_* variables do not work in that scenario and may lead to
confusion when getting started with this feature.